fix: fixes from review
